TTS/requirements.txt
f5-tts>=1.1.7
soundfile>=0.13.1
ruaccent>=1.5.8.3
torch==2.4.1
torchaudio==2.4.1
huggingface-hub[cli]
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
